+++ Package.swift
@@ -3,7 +3,15 @@ import PackageDescription
 let package = Package(
     name: "GATT",
     targets: [
-        Target(name: "GATT")
+        Target(name: "GATT"),
+        Target(
+          name: "DarwinGATT",
+          dependencies: [.Target(name: "GATT")]
+        ),
+        Target(
+          name: "LinuxGATT",
+          dependencies: [.Target(name: "GATT")]
+        )
     ],
     dependencies: [
         .Package(url: "https://github.com/PureSwift/BluetoothLinux.git", majorVersion: 3)
