+++ Sources/DarwinGATT/DarwinCentral.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 import Bluetooth
+import GATT
 
 #if os(macOS) || os(iOS) || os(tvOS) || (os(watchOS) && swift(>=3.2))
