+++ Sources/DarwinGATT/DarwinCentralError.swift
