+++ Sources/LinuxGATT/LinuxCentral.swift
@@ -10,6 +10,7 @@ import Foundation
 import Dispatch
 import Bluetooth
 import BluetoothLinux
+import GATT
 
 @available(OSX 10.12, *)
 public final class LinuxCentral: CentralProtocol {
